modularized adminController and updated filepaths where relevant
